# fix: Improve drop cache performance