## Which problem is this PR solving?

- The trace drop cache was causing cascading failures under high load when stress relief triggered. The problem was traced to the fact that the cuckoofilter library was internally locking the system random number generator repeatedly and the time spent locking and unlocking was significant. 

A test was done with a private fork of the library to remove this lock by using a different way to generate random values. This cut the time spent in the cuckoofilter by half, but this probably wouldn't have been enough to matter.

Instead, this change decouples the Add call for the drop cache; Add now simply puts an item into a large (10K item) channel that is monitored by a separate goroutine, which can in turn batch up to 5 calls to the cuckoo cache in one lock. This results in a serious performance improvement:

```
BenchmarkCuckooTraceChecker_Add-10                 	     100	      7057 ns/op
BenchmarkCuckooTraceChecker_AddParallel-10         	     100	      7385 ns/op
BenchmarkCuckooTraceChecker_Check-10               	     100	        15.83 ns/op
BenchmarkCuckooTraceChecker_CheckParallel-10       	     100	       172.9 ns/op
BenchmarkCuckooTraceChecker_CheckAddParallel-10    	     100	      7548 ns/op
```

## Short description of the changes

-