@TylerHelmuth 2023-07-06
We tested this change over the last 24 hours and feel confident that it fixes the problem.  Will merge in for 2.0.